fix: separate spa.js
