+++ english-for-kids/src/scripts/header/ctrl.js
@@ -1,37 +1,39 @@
-import { menu } from './view';
+import { menu, headerBurger, headerSwitch } from './view';
 
-const switchBtn = document.querySelector('div.header__switch');
-const burger = document.querySelector('div.header__burger');
-const menuPart = document.querySelectorAll('div.menu__part');
+// const switchBtn = document.querySelector('div.header__switch');
+// const burger = document.querySelector('div.header__burger');
 
 function switchMode() {
-  switchBtn.classList.toggle('checked');
+  headerSwitch.classList.toggle('checked');
 }
 
 function openMenu() {
-  burger.classList.toggle('header__menu-opened');
+  headerBurger.classList.toggle('header__menu-opened');
   menu.classList.toggle('menu_opened');
 }
 
 function closeMenu(e) {
   if (e.code !== 'Escape') return;
-  burger.classList.remove('header__menu-opened');
+  headerBurger.classList.remove('header__menu-opened');
   menu.classList.remove('menu_opened');
 }
 
 function closeMenuWithoutEvent() {
-  burger.classList.remove('header__menu-opened');
+  headerBurger.classList.remove('header__menu-opened');
   menu.classList.remove('menu_opened');
 }
 
 function chooseMenuPart(e) {
+  const menuPart = document.querySelectorAll('div.menu__part');
   menuPart.forEach((el) => el.classList.remove('menu__part_choosed'));
   e.target.classList.add('menu__part_choosed');
 }
 
-switchBtn.addEventListener('click', switchMode);
-burger.addEventListener('click', openMenu);
-document.addEventListener('keydown', closeMenu);
-menuPart.forEach((el) => el.addEventListener('click', chooseMenuPart));
+// headerSwitch.addEventListener('click', switchMode);
+// headerBurger.addEventListener('click', openMenu);
+// document.addEventListener('keydown', closeMenu);
+// menuPart.forEach((el) => el.addEventListener('click', chooseMenuPart));
 
-export { closeMenuWithoutEvent };
+export {
+  closeMenuWithoutEvent, openMenu, closeMenu, switchMode, chooseMenuPart,
+};
