+++ english-for-kids/src/scripts/header/view.js
@@ -1,4 +1,8 @@
 import { menuArray } from './model';
+import { openMenu, closeMenu, switchMode, chooseMenuPart } from './ctrl';
+import {
+    switchToMainPage, switchToActionAPage, switchToActionBPage, switchToActionCPage, switchToAnimalPage, switchToClothesPage
+  } from '../spa/ctrl';
 
 // !!! HEADER VIEW !!!
 // !!! HEADER VIEW !!!
@@ -19,6 +23,8 @@ headerBurger.innerHTML = `
         <line x1="4" y1="50" x2="76" y2="50"></line>
     </svg>
 `;
+headerBurger.addEventListener("click", openMenu);
+document.addEventListener('keydown', closeMenu);
 
 const h1 = document.createElement('h1');
 h1.textContent = 'English for kids';
@@ -31,6 +37,7 @@ headerSwitch.innerHTML = `
     <div class="header__switch-round"></div>
     <span>play</span>
 `;
+headerSwitch.addEventListener("click", switchMode);
 
 headerContainer.appendChild(headerBurger);
 headerContainer.appendChild(h1);
@@ -51,10 +58,31 @@ menuArray.forEach((el, i) => {
   if (i === 0) DOMel.classList.add('menu__part_choosed');
   DOMel.id = el.id;
   DOMel.innerText = el.innerText;
+  DOMel.addEventListener('click', chooseMenuPart);
   menu.appendChild(DOMel);
 });
 
 document.body.appendChild(menu);
 menu.style.top = `${header.offsetHeight}px`;
 
-export { header, menu };
+
+document.querySelector('h1').addEventListener('click', switchToMainPage);
+document.getElementById('main_page_btn').addEventListener('click', switchToMainPage);
+document.getElementById('actiona_page_btn').addEventListener('click', switchToActionAPage);
+document.getElementById('actionb_page_btn').addEventListener('click', switchToActionBPage);
+document.getElementById('actionc_page_btn').addEventListener('click', switchToActionCPage);
+document.getElementById('animala_page_btn').addEventListener('click', switchToAnimalPage);
+document.getElementById('clothes_page_btn').addEventListener('click', switchToClothesPage);
+
+document.getElementById('main_page_btn').classList.add('menu__part_onclick');
+document.getElementById('actiona_page_btn').classList.add('menu__part_onclick');
+document.getElementById('actionb_page_btn').classList.add('menu__part_onclick');
+document.getElementById('actionc_page_btn').classList.add('menu__part_onclick');
+
+document.getElementById('animala_page_btn').classList.add('menu__part_onclick');
+document.getElementById('clothes_page_btn').classList.add('menu__part_onclick');
+
+
+export {
+  header, menu, headerBurger, headerSwitch,
+};
